web/requirements.txt
requests
PyYAML
fastapi
uvicorn
psycopg2-binary
jinja2